# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhisperBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
        "409":
          description: Username already taken
  /users/check-username:
    get:
      operationId: checkUsername
      tags: [users]
      summary: Check if a username is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailability"
        "400":
          description: Invalid username
  /users/{username}:
    get:
      operationId: getPublicProfile
      tags: [users]
      summary: Get public profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UsernameAvailability:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isPremium:
          type: boolean
        defaultPublicMessages:
          type: boolean
        emailNotifications:
          type: boolean
        allowReplyNotif:
          type: boolean
        linkOpens:
          type: integer
        socialInstagram:
          type: string
          nullable: true
        socialTiktok:
          type: string
          nullable: true
        socialX:
          type: string
          nullable: true
        socialFacebook:
          type: string
          nullable: true
        socialGithub:
          type: string
          nullable: true
        socialLinkedin:
          type: string
          nullable: true
        hasSetUsername:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - username
        - isPremium
        - defaultPublicMessages
        - emailNotifications
        - allowReplyNotif
        - linkOpens
        - hasSetUsername
        - createdAt
    PublicProfile:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        socialInstagram:
          type: string
          nullable: true
        socialTiktok:
          type: string
          nullable: true
        socialX:
          type: string
          nullable: true
        socialFacebook:
          type: string
          nullable: true
        socialGithub:
          type: string
          nullable: true
        socialLinkedin:
          type: string
          nullable: true
        isPremium:
          type: boolean
        allowReplyNotif:
          type: boolean
        publicMessages:
          type: array
          items:
            $ref: "#/components/schemas/PublicMessage"
      required:
        - username
        - isPremium
        - allowReplyNotif
        - publicMessages
    PublicMessage:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        ownerReply:
          type: string
          nullable: true
        ownerRepliedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - content
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_]+$"
        displayName:
          type: string
          maxLength: 64
          nullable: true
        bio:
          type: string
          maxLength: 500
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        defaultPublicMessages:
          type: boolean
        emailNotifications:
          type: boolean
        allowReplyNotif:
          type: boolean
        socialInstagram:
          type: string
          nullable: true
        socialTiktok:
          type: string
          nullable: true
        socialX:
          type: string
          nullable: true
        socialFacebook:
          type: string
          nullable: true
        socialGithub:
          type: string
          nullable: true
        socialLinkedin:
          type: string
          nullable: true
...